+++ src/Core/Platform/VRChatContentPublisher.Platform.Windows/WindowsConst.cs
@@ -1,0 +1,6 @@
+namespace VRChatContentPublisher.Platform.Windows;
+
+internal static class WindowsConst
+{
+    public const string AppUserModelId = "vrchat-content-publisher-c7a27b26-8e90-4e82-9c10-b03f4a794b4c";
+}